Refactored export and import logic.
Introduced explicit model of translations and refactored the code to
first load the model (either from xml or from xml+excel) and then use a
visitor to output (the excel for export and xml for import).

Added automatic file splitting, by keeping a context column in the
exported excel with the path of the file being analyzed.

Added possibility to include translations that are only present in
non-default languages.

Removed support for advanced import features like transformations and
personalized escaping.